Making sure not to leak fake login button functionality into production
