+++ frontend/.gitignore
@@ -1,4 +1,6 @@
 node_modules
 public/assets/js/bundle.js
 public/assets/css/bundle.css
+public/assets/css/fontawesome-*
+public/assets/css/glyphicons-*
 .idea
